Fixes Bug in Forecast Timestamp
DateTime.now didn't agree with dry-types' DateTime type for reasons I didn't
feel like exploring right now. I also took this opportunity to extract the
various value objects from their "parent" class definitions.
