+++ app/models/location.rb
@@ -1,0 +1,16 @@
+class Location < Dry::Struct
+  attribute :latitude, Types::Float
+  attribute :longitude, Types::Float
+
+  def valid?
+    true
+  end
+
+  def invalid?
+    false
+  end
+
+  def coordinates
+    [latitude, longitude]
+  end
+end
