+++ app/models/null_location.rb
@@ -1,0 +1,13 @@
+class NullLocation
+  def coordinates
+    []
+  end
+
+  def valid?
+    false
+  end
+
+  def invalid?
+    true
+  end
+end
